KAS-4914 refactor service, better error handling, new templateversion
