+++ Dockerfile
@@ -1,1 +1,1 @@
-FROM semtech/mu-javascript-template:1.6.0
+FROM semtech/mu-javascript-template:1.8.0
