+++ lib/file.js
@@ -14,7 +14,7 @@ const parseSparqlResults = (data) => {
   });
 };
 
-const getFileById = async function (fileId) {
+async function getFileById(fileId) {
   const q = `
 PREFIX mu: <http://mu.semte.ch/vocabularies/core/>
 PREFIX nfo: <http://www.semanticdesktop.org/ontologies/2007/03/22/nfo#>
@@ -45,15 +45,47 @@ LIMIT 1
     if (parsedResults.length > 0) {
       return parsedResults[0];
     } else {
-      console.log('File found but without any results');
+      console.log("File found but without any results");
       throw new Error();
     }
   } catch (error) {
-    console.log('File not found', error);
+    console.log("File not found", error);
     throw new Error(`File with id:${fileId} not found`);
   }
 };
 
-module.exports = {
+async function getNotaFile(notaId) {
+  const queryString = `
+    PREFIX mu: <http://mu.semte.ch/vocabularies/core/>
+    PREFIX dossier: <https://data.vlaanderen.be/ns/dossier#>
+    PREFIX prov: <http://www.w3.org/ns/prov#>
+    PREFIX dct: <http://purl.org/dc/terms/>
+    PREFIX nfo: <http://www.semanticdesktop.org/ontologies/2007/03/22/nfo#>
+
+    SELECT DISTINCT ?uuid WHERE {
+      ?s a dossier:Stuk .
+      ?s mu:uuid ${sparqlEscapeString(notaId)} .
+      {
+          ?s prov:value ?file .
+      }
+      UNION
+      {
+          ?s prov:value/^prov:hadPrimarySource ?file .
+      }
+      ?file a nfo:FileDataObject ;
+          mu:uuid ?uuid ;
+          dct:format ?format .
+      FILTER(CONTAINS(?format, "application/pdf"))
+    }`;
+
+  const res = await query(queryString);
+  if (res.results.bindings.length) {
+    return res.results.bindings[0].uuid.value;
+  }
+  throw new Error(`File for Nota with id ${notaId} not found`);
+}
+
+export {
   getFileById,
+  getNotaFile,
 };
